+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.4.1 (2019-05-10)
+==================
+
+* Added `pLeftSemigroupalUpdate` and `pRightSemigroupalUpdate`.
+
 0.4.0 (2018-08-21)
 ==================
 
@@ -210,4 +215,3 @@
 ===
 
 First release.
-